Sped up drawing speeds
Made it so IntRect no longer generates an array of all its points for its IEnumerator and instead uses a lazy enumerator. This was bottlenecking DrawingArea.Update() since it gets File.rect and the IntRect constructor would generate the array of points.

This massively improved drawing performance, so you can now speedily draw on large images and with large brush sizes