Change Functions.Pow() algorithm to avoid unnecessary overflow exceptions
The previous algorithm, when computing n ^ exponent, could give an overflow exception even if n ^ exponent could be represented as an int. This would occur if n ^ (exponent + 1) couldn't be represented as an int. The issue was caused by updating the value of f. This new (and simpler/obvious) algorithm means n ^ exponent will never throw an error if it can be represented as an int.